split out KeyHandlingCollectionView
